Replace f32 color math with fixed-point integer arithmetic
Drops the palette dependency in favour of a 16-bit-per-channel Rgb type
and Q0.15 fixed-point progress/easing. On Cortex-M0+ this removes the
software float emulation (__divsf3/__addsf3/__mulsf3, ~44% of .text) from
the color path entirely.

Design notes:
- Q0.15 rather than Q0.16 so that squaring in the easing curves stays
  inside a u32; every intermediate fits 32 bits, so no 64-bit multiply
  helper is linked in.
- Color interpolation branches on direction to keep the product unsigned.
- Brightness moves from f32 to u8, applied via a shift-only divide by 255.
- HSV is reimplemented with integer sector math; the hue wheel maps onto
  the full u16 range so rotation wraps without a modulo.

Behaviour is preserved: palette's Srgb was gamma-encoded and Mix was a
naive per-channel lerp, so the integer lerp is equivalent (no gamma
conversion introduced).

+++ src/command.rs
@@ -23,8 +23,8 @@ pub enum SequencerAction<D: TimeDuration, const N: usize> {
     Restart,
     /// Clear sequence and turn off LED.
     Clear,
-    /// Set brightness multiplier (0.0-1.0, clamped).
-    SetBrightness(f32),
+    /// Set brightness multiplier (`0..=255`, where 255 is full brightness).
+    SetBrightness(u8),
 }
 
 /// Command targeting a specific LED.
